90% terminado, falta algunas boludeces que si queres lo podes cambiar
