+++ README.md
@@ -1,2 +1,12 @@
 # RepasoJava
 repaso de java previo a spring
+
+### NOTAS:
+
+1. agregue los metodos de agregar un item, eliminar, mostrar la biblioteca entera ordenada por titulo, filtrar y modificar
+
+
+2. a su vez tambien, utlice comparator para ordenar el hashmap por titulo y buscar un item por su ID
+
+
+3. tambien agregue nuevas exceptions personalizadas.
