--- repasoJava/src/Repositorio/Init/Intrep.java
@@ -1,4 +1,0 @@
-package Repositorio.Init;
-
-public class Intrep {
-}
